+++ Example/Tests/API/HTTPServiceSpec.swift
@@ -40,9 +40,12 @@ class HTTPServiceSpec: QuickSpec {
                 it("should return SEError as data and nil error") {
                     let errorJson = """
                                         {
-                                          "class": "ConnectionNotFound",
-                                          "message": "Connection with id: '987' was not found.",
-                                          "documentation_url": "https://test.com"
+                                            "error":
+                                                {
+                                                    "class": "ConnectionNotFound",
+                                                    "message": "Connection with id: '987' was not found.",
+                                                    "documentation_url": "https://test.com"
+                                                }
                                         }
                                     """
                     let (data, error) = handleResponse(from: errorJson.data(using: .utf8) , error: nil, decoder: JSONDecoder())