+++ saltedge-ios-swift/Classes/API/Models/Parameters/SETransactionParams.swift
@@ -24,10 +24,12 @@
 import Foundation
 
 public struct SETransactionParams: URLEncodable, ParametersEncodable {
+    public let perPage: String
     public let accountId: String?
     public let fromId: String?
 
-    public init(accountId: String? = nil, fromId: String? = nil) {
+    public init(perPage: String = "1000", accountId: String? = nil, fromId: String? = nil) {
+        self.perPage = perPage
         self.accountId = accountId
         self.fromId = fromId
     }